+++ spec/pio/hello_spec.rb
@@ -16,6 +16,13 @@ describe Pio::Hello do
       Then { hello.xid == 0 }
       Then { hello.body.empty? }
     end
+
+    context 'with an features-request message' do
+      Given(:features_request_dump) { [1, 5, 0, 8, 0, 0, 0, 0].pack('C*') }
+
+      When(:result) { Pio::Hello.read(features_request_dump) }
+      Then { result == Failure(BinData::ValidityError) }
+    end
   end
 
   describe '.new' do