+++ CHANGELOG.md
@@ -1,11 +1,23 @@
 # Changelog
 
+## 0.6.0 (4/15/2014)
+
+### New features
+* Added new class `Pio::Echo`, `Pio::Echo::Request` and `Pio::Echo::Reply`.
+
+
+## 0.5.1 (4/15/2014)
+
+### Bugs fixed
+* Set hello message type = 0.
+
+
 ## 0.5.0 (4/14/2014)
 
+### New features
 * Added new class `Pio::Hello`.
 
 ### Misc
-
 * Added new rake task `cucumber`.
 * Added new rake task `dump_pcap`.